refactor: split the ecrsoci module into 2 smaller ones
Split the ecrsoci module into 2 smaller ones:
- soci module: for building SOCI index and initializing related
  resources for the build.
- registry module: for interacting with a remote registry.
