+++ functions/source/soci-index-generator-lambda/utils/registry/registry.go
@@ -1,0 +1,131 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package registry
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"net/http"
+	"os"
+	"regexp"
+	"strings"
+
+	"oras.land/oras-go/v2"
+	"oras.land/oras-go/v2/content/oci"
+	"oras.land/oras-go/v2/registry/remote"
+	"oras.land/oras-go/v2/registry/remote/auth"
+
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/session"
+	"github.com/aws/aws-sdk-go/service/ecr"
+
+	"github.com/aws-ia/cfn-aws-soci-index-builder/soci-index-generator-lambda/utils/log"
+	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
+)
+
+type Registry struct {
+	orasRegistry *remote.Registry
+}
+
+var RegistryNotSupportingOciArtifacts = errors.New("Registry does not support OCI artifacts")
+
+// Initialize a remote registry
+func Init(registryUrl string) (*Registry, error) {
+	orasRegistry, err := remote.NewRegistry(registryUrl)
+	if err != nil {
+		return nil, err
+	}
+	if isEcrRegistry(registryUrl) {
+		err := authorizeEcr(orasRegistry)
+		if err != nil {
+			return nil, err
+		}
+	}
+	return &Registry{orasRegistry}, nil
+}
+
+// Pull an image from the remote registry to a local OCI Store
+// imageReference can be either a digest or a tag
+func (registry *Registry) Pull(ctx context.Context, repositoryName string, ociStore oci.Store, imageReference string) (*ocispec.Descriptor, error) {
+	log.Info(ctx, "Pulling image")
+	repo, err := registry.orasRegistry.Repository(ctx, repositoryName)
+	if err != nil {
+		return nil, err
+	}
+
+	imageDescriptor, err := oras.Copy(ctx, repo, imageReference, &ociStore, imageReference, oras.DefaultCopyOptions)
+	if err != nil {
+		return nil, err
+	}
+
+	return &imageDescriptor, nil
+}
+
+// Push a OCI artifact to remote registry
+// descriptor: ocispec Descriptor of the artifact
+// ociStore: the local OCI store
+func (registry *Registry) Push(ctx context.Context, ociStore oci.Store, indexDesc ocispec.Descriptor, repositoryName string) error {
+	log.Info(ctx, "Pushing artifact")
+
+	repo, err := registry.orasRegistry.Repository(ctx, repositoryName)
+	if err != nil {
+		return err
+	}
+
+	err = oras.CopyGraph(ctx, &ociStore, repo, indexDesc, oras.DefaultCopyGraphOptions)
+	if err != nil {
+		// TODO: There might be a better way to check if a registry supporting OCI or not
+		if strings.Contains(err.Error(), "Response status code 405: unsupported: Invalid parameter at 'ImageManifest' failed to satisfy constraint: 'Invalid JSON syntax'") {
+			fmt.Printf("[WARN] Error when pushing [err: %v]", err)
+			return RegistryNotSupportingOciArtifacts
+		}
+		return err
+	}
+	return nil
+}
+
+// Check if a registry is an ECR registry
+func isEcrRegistry(registryUrl string) bool {
+	ecrRegistryUrlRegex := "\\d{12}\\.dkr\\.ecr\\.\\S+\\.amazonaws\\.com"
+	match, err := regexp.MatchString(ecrRegistryUrlRegex, registryUrl)
+	if err != nil {
+		panic(err)
+	}
+	return match
+}
+
+// Authorize ECR registry
+func authorizeEcr(ecrRegistry *remote.Registry) error {
+	// getting ecr auth token
+	input := &ecr.GetAuthorizationTokenInput{}
+	var ecrClient *ecr.ECR
+	ecrEndpoint := os.Getenv("ECR_ENDPOINT") // set this env var for custom, i.e. non default, aws ecr endpoint
+	if ecrEndpoint != "" {
+		ecrClient = ecr.New(session.New(&aws.Config{Endpoint: aws.String(ecrEndpoint)}))
+	} else {
+		ecrClient = ecr.New(session.New())
+	}
+	getAuthorizationTokenResponse, err := ecrClient.GetAuthorizationToken(input)
+	if err != nil {
+		return err
+	}
+
+	if len(getAuthorizationTokenResponse.AuthorizationData) == 0 {
+		return errors.New("Couldn't authorize with ECR: empty authorization data returned")
+	}
+
+	ecrAuthorizationToken := getAuthorizationTokenResponse.AuthorizationData[0].AuthorizationToken
+	if len(*ecrAuthorizationToken) == 0 {
+		return errors.New("Couldn't authorize with ECR: empty authorization token returned")
+	}
+
+	ecrRegistry.RepositoryOptions.Client = &auth.Client{
+		Header: http.Header{
+			"Authorization": {"Basic " + *ecrAuthorizationToken},
+			"User-Agent":    {"SOCI Index Builder (oras-go)"},
+		},
+	}
+	return nil
+}
